Update activity_course_meeting_schedule.xml
I have removed the spinners I created before and replaced them with the TimePicker widgets. Refer to the link I posted on discord to write the java code to get the proper time formats. Disregard the values file with the string arrays for the times as that file is not needed at this time.