Update activity_course_information.xml
Removed the labels(TextViews), since the editable textbox hints already give labels. I took out the text box for the invite code since we decided this will automatically be generated from the database. I was able to fix the multiline text box for the course description so that it no longer pushes down the button if a lot of text is typed in. A scroll bar will now appear if more than 3 lines of text are typed.